Add sim::MCParticleLite (GeV units) as a parent class of sim::MCMiniPart (MeV units)
