+++ lardataobj/MCBase/classes_def.xml
@@ -30,8 +30,12 @@
  <version ClassVersion="11" checksum="1998549158"/>
  <version ClassVersion="10" checksum="1862388136"/>
 </class>
-<class name="sim::MCMiniPart" ClassVersion="10" >
+<class name="sim::MCMiniPart" ClassVersion="11" >
+ <version ClassVersion="11" checksum="3396563188"/>
  <version ClassVersion="10" checksum="924559196"/>
+</class>
+<class name="sim::MCParticleLite" ClassVersion="10" >
+ <version ClassVersion="10" checksum="1468541834"/>
 </class>
 
 <class name="std::vector<sim::MCHit>"/>
@@ -58,6 +62,10 @@
 <class name="std::vector<sim::MCMiniPart>"/>
 <class name="art::Wrapper< std::vector<sim::MCMiniPart>>"/>
 <class name="art::Ptr<sim::MCMiniPart>" />
+
+<class name="std::vector<sim::MCParticleLite>"/>
+<class name="art::Wrapper< std::vector<sim::MCParticleLite>>"/>
+<class name="art::Ptr<sim::MCParticleLite>" />
 
 <ioread
   version="[10]"
